requirements.txt
Flask==0.8
Flask-Script==0.3.1
Jinja2==2.6
Werkzeug==0.8.2
argparse==1.2.1
certifi==0.0.7
chardet==1.0.1
lxml==2.3.3
pymongo==2.1.1
python-dateutil==2.0
requests==0.10.1
wsgiref==0.1.2
